Add failing test case for issue #56
https://github.com/nco/pynco/issues/56

`nco.ncatted` and other commands can hang indefinitely if an output file
already exists. The `nco` tool prompts the user for input before
overwriting the file, but the user never sees the prompt and is unable
to provide a response.

By closing stdin on the subprocess, the `nco` tool aborts after
prompting for input. This is better than an infinite hang.